Add files via upload
This is a program for analyzing text from the clipboard. Supports analysis of Russian and English text. Counts the number of vowels, consonants, the total number of characters, and the total number of words in the text. System libraries are used: sys, os, and colorama and paperclip.
This program created for terminal, and work correctly on win32 and unix like systems.